+++ src/enums/ChangeMethodEnum.ts
@@ -1,4 +1,5 @@
 enum ChangeMethodEnum {
+  GrantWritePermission = 'grant_write_permission',
   Set = 'set',
   StorageDeposit = 'storage_deposit',
   StorageWithdraw = 'storage_withdraw',