+++ src/enums/ErrorCodeEnum.ts
@@ -1,0 +1,10 @@
+enum ErrorCodeEnum {
+  // accounts
+  InvalidAccountIdError = 2000,
+  AccountNotFoundError = 2001,
+
+  // keys
+  KeyNotAllowedError = 3000,
+}
+
+export default ErrorCodeEnum;