+++ test/helpers/signAndSendTransaction/types/IOptions.ts
@@ -1,0 +1,8 @@
+import { Account, transactions } from 'near-api-js';
+
+interface IOptions {
+  signerAccount: Account;
+  transaction: transactions.Transaction;
+}
+
+export default IOptions;