+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# [1.0.0-beta.3](https://github.com/NEARBuilders/near-social-js/compare/v1.0.0-beta.2...v1.0.0-beta.3) (2024-06-12)
+
+
+### Features
+
+* implement function for grant write permission ([#16](https://github.com/NEARBuilders/near-social-js/issues/16)) ([f3047c7](https://github.com/NEARBuilders/near-social-js/commit/f3047c7a66bc5309a8f66df85f72e68e5e39bcfc))
+
+# [1.0.0-beta.2](https://github.com/NEARBuilders/near-social-js/compare/v1.0.0-beta.1...v1.0.0-beta.2) (2024-06-05)
+
+
+### Features
+
+* change blockhash and nonce parameters to be optional ([#20](https://github.com/NEARBuilders/near-social-js/issues/20)) ([7f1cf2e](https://github.com/NEARBuilders/near-social-js/commit/7f1cf2ee2e5d743a1a37bc0b5f6e774e8900e41a))
+
 # 1.0.0-beta.1 (2024-05-15)